requirements.txt
absl-py==1.3.0
addict==2.4.0
backports.functools-lru-cache==1.6.4
brotlipy==0.7.0
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
chardet==5.2.0
charset-normalizer==2.1.1
cloudpickle==2.2.1
crcmod==1.7
cryptography==38.0.4
cycler==0.11.0
cython-bbox==0.1.3
Deprecated==1.2.14
dnspython==2.3.0
filelock==3.8.0
fonttools==4.37.4
google-auth==2.12.0
google-auth-oauthlib==0.4.6
grpcio==1.57.0
huggingface-hub==0.10.1
idna==3.4
importlib-metadata==4.11.4
ipaddress==1.0.23
joblib==1.2.0
kiwisolver==1.4.4
lap==0.4.0
loguru==0.6.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib
mkl-fft
mkl-random
mkl-service
motmetrics==1.2.5
msgpack==1.0.5
# MultiScaleDeformableAttention==1.0
numpy
nvidia-ml-py3==7.352.0
oauthlib==3.2.1
opencv-python==4.6.0.66
packaging==23.1
pandas==1.3.5
Pillow==9.2.0
pip==23.2.1
portalocker==2.5.1
prettytable==3.4.1
protobuf
py-spy==0.3.14
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.5
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-etcd==0.4.5
pytz==2022.4
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
schedule==1.2.0
scikit-learn==1.0.2
scipy==1.7.3
setuptools==65.5.0
six==1.16.0
sklearn==0.0
tensorboard
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.0.1
threadpoolctl==3.1.0
timm==0.6.11
torch==1.12.1
torch-tb-profiler==0.4.0
torchaudio==0.12.1
torchvision==0.13.1
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.11
wcwidth==0.2.5
Werkzeug==2.2.2
wheel==0.37.1
wrapt==1.15.0
xmltodict==0.13.0
yapf==0.32.0
zipp==3.9.0
